[Refactor] 공고 패키지 메서드 및 응답 위치 변경
