+++ backend/src/main/java/com/sabujaks/irs/domain/announcement/service/AnnouncementService.java
@@ -16,6 +16,7 @@ import com.sabujaks.irs.domain.auth.repository.RecruiterRepository;
 import com.sabujaks.irs.domain.company.model.entity.Company;
 import com.sabujaks.irs.domain.company.repository.CompanyBenefitsRepository;
 import com.sabujaks.irs.domain.company.repository.CompanyRepository;
+import com.sabujaks.irs.domain.company.service.CompanyService;
 import com.sabujaks.irs.domain.system.model.entity.BaseInfo;
 import com.sabujaks.irs.domain.system.repository.BaseInfoRepository;
 import com.sabujaks.irs.domain.resume.repository.ResumeRepository;
@@ -43,9 +44,10 @@ public class AnnouncementService {
     private final CustomFormRepository customFormRepository;
     private final CustomLetterFormRepository letterFormRepository;
     private final BaseInfoRepository baseInfoRepository;
-    private final CompanyBenefitsRepository companyBenefitsRepository;
     private final CompanyRepository companyRepository;
     private final ResumeRepository resumeRepository;
+    private final CompanyService companyService;
+
 
     /*******채용담당자 공고 등록 (step1)***********/
     @Transactional
@@ -127,13 +129,13 @@ public class AnnouncementService {
 
         Long recruiterIdx = customUserDetails.getIdx();
         Optional<Recruiter> resultRecruiter = recruiterRepository.findByRecruiterIdx(recruiterIdx);
-        if(!resultRecruiter.isPresent()) {
+        if(resultRecruiter.isEmpty()) {
             throw new BaseException(BaseResponseMessage.ANNOUNCEMENT_REGISTER_STEP_ONE_FAIL_NOT_RECRUITER);
         }
 
         // 예외 처리) 공고가 잘 저장되어 있는지 먼저 확인 필요
         Optional<Announcement> resultAnnouncement = announcementRepository.findByAnnounceIdx(dto.getAnnouncementIdx());
-        if (!resultAnnouncement.isPresent()) {
+        if (resultAnnouncement.isEmpty()) {
             throw new BaseException(BaseResponseMessage.ANNOUNCEMENT_REGISTER_STEP_TWO_FAIL_NOT_FOUND);
         }
 
@@ -185,96 +187,6 @@ public class AnnouncementService {
                 .descriptionList(emsiDes)
                 .customFormLetterIdList(emsiLetterIdx)
                 .build();
-    }
-
-
-    /*******공고 등록 페이지 클릭시 채용담당자 정보 조회***********/
-    public RecruiterInfoReadRes readRecruiterInfo(CustomUserDetails customUserDetails) throws BaseException {
-        Long recruiterIdx = customUserDetails.getIdx();
-        // 채용담당자 확인
-        Optional<Recruiter> resultRecruiter = recruiterRepository.findByRecruiterIdx(recruiterIdx);
-        if(resultRecruiter.isPresent()) {
-            String phoneNumber = resultRecruiter.get().getPhoneNumber();
-
-            String phoneNumberDasi = phoneNumber.replace("-", "");
-
-            String phone1 = phoneNumberDasi.substring(0, 3);
-            String phone2 = phoneNumberDasi.substring(3, 7);
-            String phone3 = phoneNumberDasi.substring(7);
-
-            return RecruiterInfoReadRes.builder()
-                    .managerName(resultRecruiter.get().getName())
-                    .managerContact(phoneNumber)  // 전체 전화번호를 유지하는 경우
-                    .phone1(phone1)               // 첫 번째 부분
-                    .phone2(phone2)               // 두 번째 부분
-                    .phone3(phone3)               // 세 번째 부분
-                    .managerEmail(resultRecruiter.get().getEmail())
-                    .build();
-        } else {
-            // 채용담당자 유저에서 찾을 수 없을 때
-            throw new BaseException(BaseResponseMessage.ANNOUNCEMENT_REGISTER_STEP_ONE_FAIL_NOT_RECRUITER);
-        }
-    }
-
-
-    /*******공고 등록 페이지 클릭시 채용담당자 기업 복리후생 조회***********/
-    public CompanyInfoReadRes readCompanyInfo(String recruiterEmail) throws BaseException {
-        // 채용담당자 확인
-        Optional<Recruiter> resultRecruiter = recruiterRepository.findByRecruiterEmail(recruiterEmail);
-        if(resultRecruiter.isPresent()) {
-            Recruiter recruiter = resultRecruiter.get();
-            // 채용담당자가 등록한 기업 확인
-            Optional<Company> resultCompany = companyRepository.findByRecruiterIdx(recruiter.getIdx());
-            if (resultCompany.isPresent()) {
-                Long companyIdx = resultCompany.get().getIdx();
-
-                // 기업 복리후생 테이블에서 해당 기업의 복리후생 코드들을 가져옴
-                // 스트림 - 리스트 내의 각 요소에 대해 반복작업 쉽게 가능
-                List<String> companyBenefitsCodes = companyBenefitsRepository
-                        .findAllByCompanyIdx(companyIdx)
-                        .stream()
-                        .map(CompanyBenefits::getCode)
-                        .collect(Collectors.toList());
-
-                // 기준 코드 테이블에서 가져온 복리후생 코드를 조회
-                List<BaseInfo> benefitsBaseInfos = baseInfoRepository.findByCodeIn(companyBenefitsCodes);
-
-                // 대분류와 소분류를 저장할 맵
-                Map<String, List<String>> benefitsMap = new HashMap<>();
-
-                // 기준 코드 테이블에서 응답 생성
-                for (BaseInfo baseInfo : benefitsBaseInfos) {
-                    String category = baseInfo.getDescription();
-                    String parentCode = baseInfo.getParentCode();
-
-                    if (parentCode == null) {
-                        // 부모 코드가 없으면 대분류로 처리
-                        benefitsMap.putIfAbsent(category, new ArrayList<>());
-                    } else {
-                        // 부모 코드가 있으면 대분류의 소분류에 추가
-                        String parentCategory = baseInfoRepository.findByCode(parentCode).getDescription();
-                        benefitsMap.computeIfAbsent(parentCategory, k -> new ArrayList<>()).add(category);
-                    }
-                }
-
-                // 응답 생성
-                // .entrySet() : Map의 각 항목을 Map.Entry 객체(키-값 쌍)로 반환
-                List<BenefitCategory> benefitCategories = benefitsMap.entrySet().stream()
-                        .map(entry -> new BenefitCategory(entry.getKey(), entry.getValue()))
-                        .collect(Collectors.toList());
-
-                // 최종 응답 반환
-                return CompanyInfoReadRes.builder()
-                        .benefitsDataList(benefitCategories)
-                        .build();
-            } else {
-                // 기업을 찾을 수 없을 때
-                throw new BaseException(BaseResponseMessage.COMPANY_INFO_FAIL_NOT_REGISTER);
-            }
-        } else {
-            // 채용담당자 유저에서 찾을 수 없을 때
-            throw new BaseException(BaseResponseMessage.ANNOUNCEMENT_REGISTER_STEP_ONE_FAIL_NOT_RECRUITER);
-        }
     }
 
 
@@ -351,7 +263,7 @@ public class AnnouncementService {
                         .companyBusiness(company.getBusiness())
                         .companyUrl(company.getUrl())
                         .companyAddress(company.getAddress())
-                        .companyBenefitsDataList(readCompanyInfo(resultRecruiter.getEmail()).getBenefitsDataList())
+                        .companyBenefitsDataList(companyService.readCompanyInfo(resultRecruiter.getEmail()).getBenefitsDataList())
                         .build();
 
                 return announcementReadDetailRes;
