+++ backend/src/main/java/com/sabujaks/irs/domain/auth/model/response/RecruiterInfoReadRes.java
@@ -1,4 +1,4 @@
-package com.sabujaks.irs.domain.announcement.model.response;
+package com.sabujaks.irs.domain.auth.model.response;
 
 import lombok.*;
