+++ backend/src/main/java/com/sabujaks/irs/domain/company/controller/CompanyController.java
@@ -1,6 +1,6 @@
 package com.sabujaks.irs.domain.company.controller;
 
-import com.sabujaks.irs.domain.announcement.model.request.CustomFormCreateReq;
+import com.sabujaks.irs.domain.company.model.response.CompanyInfoReadRes;
 import com.sabujaks.irs.domain.company.model.request.CompanyCreateReq;
 import com.sabujaks.irs.domain.company.model.response.CompanyCreateRes;
 import com.sabujaks.irs.domain.company.model.response.CompanyReadRes;
@@ -16,7 +16,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -25,24 +24,6 @@ import java.util.List;
 public class CompanyController {
     private final CompanyService companyService;
     private final CloudFileUpload cloudFileUpload;
-
-    // 기업 정보 조회
-    @GetMapping("/read")
-    public ResponseEntity<BaseResponse<CompanyReadRes>> read(
-        @AuthenticationPrincipal CustomUserDetails customUserDetail) throws BaseException {
-
-        CompanyReadRes response = companyService.readCompany(customUserDetail);
-        return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS_REGISTER, response));
-    }
-
-    // 기업 이미지 조회
-    @GetMapping("/read-img")
-    public ResponseEntity<BaseResponse<CompanyReadRes>> readImg(Long companyIdx) throws BaseException {
-
-        CompanyReadRes response = companyService.readCompanyImg(companyIdx);
-        return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS_REGISTER, response));
-    }
-
 
     // 기업 정보 등록
     @PostMapping("/create")
@@ -60,4 +41,30 @@ public class CompanyController {
 
         return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS, response));
     }
+
+    // 기업 복리후생 조회 (공고 등록을 위한 조회2)
+    @GetMapping("/read-company-info")
+    public ResponseEntity<BaseResponse<CompanyInfoReadRes>> readCompanyInfo(String recruiterEmail) throws BaseException {
+
+        CompanyInfoReadRes response = companyService.readCompanyInfo(recruiterEmail);
+        return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS_REGISTER, response));
+    }
+
+    // 기업 정보 조회
+    @GetMapping("/read")
+    public ResponseEntity<BaseResponse<CompanyReadRes>> read(
+            @AuthenticationPrincipal CustomUserDetails customUserDetail) throws BaseException {
+
+        CompanyReadRes response = companyService.readCompany(customUserDetail);
+        return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS_REGISTER, response));
+    }
+
+    // 기업 이미지 조회
+    @GetMapping("/read-img")
+    public ResponseEntity<BaseResponse<CompanyReadRes>> readImg(Long companyIdx) throws BaseException {
+
+        CompanyReadRes response = companyService.readCompanyImg(companyIdx);
+        return ResponseEntity.ok(new BaseResponse(BaseResponseMessage.COMPANY_INFO_SUCCESS_REGISTER, response));
+    }
+
 }
